Fixed some defaults in the header comments.
Renamed and relogic'd -reuseksk and -reusezsk to -genksk and -genzsk.  This
makes key re-use the default.
Added -genkeys.

Made the display of the random-number-generator-related entropy message depend
on a config option.  The default is to display.
Made the entropy message shorter so that it's more likely to be read and
to be useful.

Modified to write the keyrec file after each set of keys (KSK and ZSKs) are
generated.

Deleted a few unused subroutine arguments.

Added a message saying when the zone expires and insisting that the keys not
be deleted until then.  These messages are likely to be honed further.

Adjusted usage message and pod as needed for the above changes.