Added external interfaces for rollmgr_droppid(). Added Darwin as a supported O/S.
Modified existing unix_droppid() to do a "ps -ax" and then munge about in the
output.  We *should* have been able to do open(FOO,"ps -p $rdpid |") and skip
the search loop.  However, the $rdpid seems to be dropped by the open() and
thus this hard search was required.


git-svn-id: https://www.dnssec-tools.org/svn/dnssec-tools/trunk@1199 c4029f9c-b713-0410-a448-fae9e31a314a