Added support for -rollall and -rollzone commands.
In skipnow(), commented out the setting of the kskphase and zskphase rollrec
fields to 0.  When a zone was marked as a skip zone (by rollctl), these lines
were forcing the zone to drop out of whatever rollover state they were in.
This is probably not the right thing to do.  (The "probably" is why the lines
were commented out, rather than being deleted.)