Added a column for zonegroups and allowed for hiding/showing the zone groups column.  Also did proper enabling/disabling of the "Show All Keysets" menu items.
Deleted most of the text from the help menu item.  This was copied from
the pod and reformatted for Tk.  Rather than letting the one get out of
day, which was all too often the case, the remaining text is relatively
non-specific and refers the reader to the pod.

Added a paint flag parameter to disp_menus().  This allows for only
repainting the display during display-menu commands when it's actually
needed, rather than having several choppy updates happening when a
single menu command is actually handled by multiple calls to disp_menus().

Updated the pod to cover the current set of menu commands.
 


git-svn-id: https://www.dnssec-tools.org/svn/dnssec-tools/trunk@6118 c4029f9c-b713-0410-a448-fae9e31a314a