At startup, rollerd marks the domains as being under its control.
Added support for the signzone command.  The command may be sent by either
zonesigner or rolctl.  This causes zonesigner to be executed with the
-signonly option, in order to sign the zonefile without generating any new
keys or causing any rollover actions to occur.

The "-rollmgr rollerd" option is added to all zonesigner executions.

The pod description was rewritten to make this information more useful.

Fixed a typo.