Fix build on musl libc
Function such as ns_name_pton are glibc specific and are not available
on other libc's like musl. This patch should make it possible to use
those internal function. The need for this patch is due to some of the
structs like __ns_sect are already defined on musl and glibc, thus would
result in an redefinition error.

First discovered on Gentoo linux.

Bug: https://bugs.gentoo.org/897862
Signed-off-by: Brahmajit Das <brahmajit.xyz@gmail.com>